继续完善虚拟纹理的逻辑

+++ src/9.other/9.10.VirtualTexture/TextureAtlas.h
@@ -22,6 +22,6 @@ public:
 
 private:
 	VirtualTextureInfo* m_info;
-	unsigned short m_texture;
+	GLuint m_texture;
 	StagingPool m_stagingPool;
 };
